refactor: move present footer to component
